requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.2.1
cryptography==46.0.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
numpy==2.3.3
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.1.1
scikit-learn==1.7.2
scipy==1.16.2
SQLAlchemy==2.0.43
threadpoolctl==3.6.0
typing_extensions==4.15.0
Werkzeug==3.1.3
